+++ de.tbonsack.karpfediem.pokemon.cardmanager.model/src/de/tbonsack/karpfediem/pokemon/cardmanager/model/services/PreferenceService.java
@@ -2,5 +2,5 @@ package de.tbonsack.karpfediem.pokemon.cardmanager.model.services;
 
 public interface PreferenceService {
 
-	public long getNextID();
+	public int getNextID();
 }
